Fix dz file seeking on MSVC
The fix for 2-4GB files seems to break things on the modern Microsoft
compiler. Seems quite hacky to use a WinAPI function for seeking but
then read with a C standard library function? Using normal fseek for
now. If it does not work for 2-4GB files now so be it, at least it works
at all for normal files.